Allow `sympy` types as `MVector.subtype`, etc.
Treat various `sympy` types as "numbers", so they can be used as
multi-vector coefficients. Improve pytest-s.
Note: `sympy.Symbol`-s are not fully supported, because of the string
representation issues.

Avoid overwrite of `MVector.value by `astype()` method, this was
happening when `astype(sympy.Number)` is called when `dtype==object`.